+++ utils/sys_err.cpp
@@ -1,0 +1,4 @@
+#include "webserv.hpp"
+
+int	sys_err(const std::string& msg)
+{std::cerr << msg << std::string(strerror(errno)) << std::endl; return -1;}
